# swagger/swagger.yaml
basePath: /api/v1
definitions:
  report.CreateReportRequest:
    properties:
      country:
        type: string
      description:
        maxLength: 2000
        type: string
      event_type:
        enum:
        - armed_conflict
        - use_of_force
        - explosion
        - terrorism
        - civil_unrest
        - other
        type: string
      latitude:
        type: number
      location_name:
        type: string
      longitude:
        type: number
      severity:
        enum:
        - critical
        - high
        - medium
        - low
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - country
    - event_type
    - latitude
    - longitude
    - severity
    - title
    type: object
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.RegisterRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - display_name
    - email
    - password
    type: object
  user.UpdateProfileRequest:
    properties:
      display_name:
        type: string
    required:
    - display_name
    type: object
host: localhost:8080
info:
  contact: {}
  description: Global Threat Report API — aggregates conflict events from GDELT and
    user reports.
  title: Conflux API
  version: "1.0"
paths:
  /admin/sync/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get GDELT sync status
      tags:
      - admin
  /admin/sync/trigger:
    post:
      description: Runs a full GDELT sync cycle synchronously and returns the result
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Trigger GDELT sync manually
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login and get access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /events:
    get:
      description: Returns paginated, filterable list of conflict events (GDELT +
        user reports)
      parameters:
      - description: 'Comma-separated: critical,high,medium,low'
        in: query
        name: severity
        type: string
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by country code
        in: query
        name: country
        type: string
      - description: gdelt or user_report
        in: query
        name: source
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: 'Bounding box: min_lng,min_lat,max_lng,max_lat'
        in: query
        name: bbox
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: date_desc
        description: date_desc, date_asc, severity
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List events
      tags:
      - events
  /events/{id}:
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get event by ID
      tags:
      - events
  /events/nearby:
    get:
      description: Returns events within a radius of a geographic point
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lng
        required: true
        type: number
      - default: 50
        description: Radius in km (max 500)
        in: query
        name: radius_km
        type: number
      - description: Filter by severity
        in: query
        name: severity
        type: string
      - default: 20
        description: Max results (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Find nearby events
      tags:
      - events
  /reports:
    post:
      consumes:
      - application/json
      description: Creates a threat event report and auto-clusters with nearby reports
      parameters:
      - description: Report data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/report.CreateReportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Submit a user report
      tags:
      - reports
  /reports/{id}:
    delete:
      description: Soft-deletes a report (only the owner can delete)
      parameters:
      - description: Report/Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete my report
      tags:
      - reports
  /reports/me:
    get:
      description: Returns paginated list of the authenticated user's reports
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List my reports
      tags:
      - reports
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"